mobilenet/requirements.txt
azure-functions
onnxruntime
numpy
Pillow
onnx
line-profiler